requirements.txt
fastapi
uvicorn
torch>=2.0.0
redis
pydantic
prometheus_client
# milvus
numpy<2.0.0
faiss-gpu
hdbscan
pyjwt
kubernetes
httpx

# LoRA/PEFT関連
peft
accelerate

pymilvus

# PostgreSQL 接続用ライブラリ (例)
psycopg2-binary

# 省略可能な既存パッケージ類はコメント扱い
# ...
# ...
# そのほか既存の記述は省略